upgrade dependencies and added width parameter

+++ lib/features/image_carousel/presentation/ui/widget/m_dot_indicator.dart
@@ -3,7 +3,7 @@ import 'package:dots_indicator/dots_indicator.dart';
 
 class MDotIndicator extends StatelessWidget {
   final int count;
-  final double position;
+  final int position;
   final Color dotColor;
 
   const MDotIndicator({
